左侧去重

+++ src/main/java/com/juyo/visa/admin/company/service/CompanyViewService.java
@@ -297,7 +297,9 @@ public class CompanyViewService extends BaseService<TCompanyEntity> {
 					dbDao.insert(entityNew);
 				}
 			}else {
-				dbDao.delete(entity);
+				if(!Util.isEmpty(entity)) {
+					dbDao.delete(entity);
+				}
 			}
 		}
